Adds VisualStudio project structure, moves files into sub-folder
Also adds a few more test cases to the project functions.
