+++ Lab-3/Lab-3.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29411.108
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lab-3", "Lab-3.vcxproj", "{CF699BFA-325D-44FE-8C33-CA3146972E5C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Debug|x64.ActiveCfg = Debug|x64
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Debug|x64.Build.0 = Debug|x64
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Debug|x86.ActiveCfg = Debug|Win32
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Debug|x86.Build.0 = Debug|Win32
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Release|x64.ActiveCfg = Release|x64
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Release|x64.Build.0 = Release|x64
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Release|x86.ActiveCfg = Release|Win32
+		{CF699BFA-325D-44FE-8C33-CA3146972E5C}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4EB060DE-2C8B-4EE4-8146-CA9F87EA6A6E}
+	EndGlobalSection
+EndGlobal
